Refactor utility module (#490)
