+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 - `create_cs_from_xyz`, `create_cs_from_xy_and_orientation`, `create_cs_from_yz_and_orientation` and
   `create_cs_from_xz_and_orientation` from `CoordinateSystemManager`. Use `create_cs_from_axis_vectors`
   instead. [[#472]](https://github.com/BAMWelDX/weldx/pulls/472)
+- `is_column_in_matrix`, `is_row_in_matrix`, `to_float_array`, `to_list`, `matrix_is_close`, `vector_is_close`
+  and `triangulate_geometry` from `weldx.util` [[#490]](https://github.com/BAMWelDX/weldx/pull/490)
 
 ### changes
 
@@ -39,6 +41,8 @@
   , ``interp_time``) [[#476]](https://github.com/BAMWelDX/weldx/pull/476)
 - `WeldxAccessor.time_ref` setter now raises a `TypeError` if `None` is passed to it
   [[#489]](https://github.com/BAMWelDX/weldx/pull/489)
+- move xarray related utility functions into `weldx.util.xarray` and all other ones into `weldx.util.util`. Content from
+  both submodules can still be accessed using `weldx.util` [[#490]](https://github.com/BAMWelDX/weldx/pull/490)
 
 ### fixes
