+++ devtools/scripts/clean_notebooks.py
@@ -2,7 +2,7 @@
 if __name__ == "__main__":
     from pathlib import Path
 
-    from weldx.util import _clean_notebook
+    from weldx.util.util import _clean_notebook
 
     p = Path(__file__).resolve().parent
     notebooks = p.glob("../../tutorials/*.ipynb")
