+++ weldx/util/__init__.py
@@ -1,0 +1,4 @@
+"""Contains utility functions and classes."""
+
+from .util import *
+from .xarray import *
